한국국제협력단 협력국 개발지표 및 ODA 지원 실적

한국국제협력단(KOICA)에서 제공하는 협력국(총 138개국/ 사무소 주재국 44개, 겸임국 94개)의 개발 지표 및 ODA 지원 실적. *개발지표: 인구, 경제, 인간개발, 빈곤, 교육, 재정, 거버넌스, 비즈니스, 산업인프라, 성평등 *출처: World Bank Open data, Human Development Data Center, Fragile States Index, Transparency International, UN Department of Economic and Social Affairs Public Institutions, OECD Statistics, KOICA ODA 정보센터

사용법

odaIndex

양식

A list with 80 data frame.

WB_국가명

WB 국가명 WB: World Bank 세계은행

ISO3_국가코드

ISO3 국가코드 ISO : International Organization for Standardization

DAC_국가코드

DAC 국가코드 DAC : Development Assistance Committee, 개발 원조 위원회

UNDP_국가명

UNDP 국가명 UNDP : 개도국에 대한 유엔의 개발 원조계획을 조정하기 위한 기관

WB_지역

WB 지역 WB: World Bank 세계은행

KOICA_국가명_국문

KOICA 국가명 국문

KOICA_지역명_국문

KOICA 지역명 국문

KOICA_지역명_영문

KOICA 지역명 영문

KOICA_지역명_부서

KOICA 지역명 부서

국가협력전략.대상국가

국가협력전략 대상 국가 여부 yes, no

KOICA.사무소.주재.여부

KOICA 사무소 주재 여부 yes, no

겸임국.관할.사무소명

겸임국 관할 사무소명

OECD_국가명

OECD 국가명

OECD소득.기준

OECD소득 기준

인구수_기준연도

해당 인구수 데이터의 기준연도 2020, 2011

인구수

인구수

GDP_기준연도

해당 GDP 데이터의 기준연도

X1인당.국내총생산

1인당.국내총생산

국민총생산.기준연도

해당 국민총생산 데이터의 기준연도

X1인당.국민총생산

1인당 국민총생산

빈곤율.기준연도

해당 빈곤율 데이터의 기준연도

빈곤율

빈곤율

지니계수.기준연도

해당 지니계수 데이터의 기준연도

지니계수

지니계수 지니계수 : 빈부격차와 계층간 소득의 불균형 정도를 나타내는 수치로, 소득이 어느 정도 균등하게 분배되는지를 알려줌. 지니계수는 0부터 1까지의 수치로 표현되는데, 값이 ‘0’(완전평등)에 가까울수록 평등하고 ‘1’(완전불평등)에 근접할수록 불평등하다는 것을 나타냄

문해율.기준연도

해당 문해율 데이터의 기준연도

문해율

문해율

중등학교.등록율.여학생..기준연도

해당 여학생 중등학교 등록율 데이터의 기준연도

중등학교.등록율.여학생.

여학생 중등학교 등록율

중등학교.등록률.남학생..기준연도

해당 남학생 중등학교 등록율 데이터의 기준연도

중등학교.등록율.남학생.

남학생 중등학교 등록율

중등학교.등록율.젠더.패리티.인덱스.기준연도

중등학교 등록율 젠더 패리티 인덱스 데이터의 기준연도

중등학교.등록율.젠더.패리티.인덱스

중등학교 등록율 젠더 패리티 인덱스

세금.기준연도

세금 데이터의 기준연도

세금

세금

ODA.순.수원액.기준연도

ODA 순 수원액 데이터의 기준연도

ODA.순.수원액

ODA 순 수원액

기업여건.기준연도

기업여건 데이터의 기준연도

기업여건

기업여건

해외직접투자.기준연도

해외직접투자 데이터의 기준연도

해외직접투자

해외직접투자

X100명당.이동통신.사용률.기준연도

100명당 이동통신 사용률 데이터의 기준연도

X100명당.이동통신.사용률

100명당 이동통신 사용률

X1인당.전기.소비량.기준연도

1인당 전기 소비량 데이터의 기준연도

X1인당.전기.소비량

1인당 전기 소비량

인터넷.사용률.기준연도

인터넷 사용률 데이터의 기준연도

인터넷.사용률

인터넷 사용률

취약국가.지수.기준연도

취약국가 지수 데이터의 기준연도

취약국가.지수

취약국가 지수

취약국가.순위

취약국가 순위

부패인식지수.기준연도

부패인식지수 데이터의 기준연도

부패인식점수

부패인식점수

부패인식순위

부패인식순위

전자정부.지수.기준연도

전자정부 지수 데이터의 기준연도

전자정부.지수.순위

전자정부 지수 순위

전자정부지수

전자정부지수

인간개발지수.기준연도

인간개발지수 데이터의 기준연도

인간개발지수순위

인간개발지수순위

인간개발지수

인간개발지수

젠더개발지수.기준연도

젠더개발지수 데이터의 기준연도

젠더개발지수

젠더개발지수

젠더.불평등.지수.기준연도

젠더 불평등 지수 데이터의 기준연도

젠더.불평등.지수

젠더 불평등 지수

여성국회의원.비율.기준연도

여성국회의원비율 데이터의 기준연도

여성국회의원.비율

여성국회의원 비율

여성실업률.기준연도

여성실업률 데이터의 기준연도

여성실업률

여성실업률

ODA.규모.기준연도

ODA 규모 데이터의 기준연도

수원국.ODA.규모

수원국 ODA 규모

양자.지원.규모

양자 지원 규모

GNI.대비.ODA.비율

GNI 대비 ODA 비율

한국.ODA.규모.기준연도

한국 ODA 규모 데이터의 기준연도

한국.ODA.규모

한국 ODA 규모

KOICA.지원.규모_공공행정

KOICA 지원 공공행정 분야 규모

KOICA.지원.규모_교육

KOICA 지원 교육 분야 규모

KOICA.지원.규모_기술환경에너지

KOICA 지원 기술환경에너지 분야 규모

KOICA.지원.규모_기타

KOICA 지원 기타 분야 규모

KOICA.지원.규모_긴급구호

KOICA 지원 긴급구호 분야 규모

KOICA.지원.규모_농림수산

KOICA 지원 농림수산 규모

KOICA.지원.규모_보건

KOICA 지원 보건 규모

KOICA.지원.규모_전체

KOICA 지원 전체 규모

소스코드

https://www.data.go.kr/data/15088160/fileData.do

예시 코드

DT::datatable(head(odaIndex,50))